Update resources, types, and UI/toolbar setup
Multiple updates to improve resource handling, 64-bit/size-safety, and UI setup:

- Initialize application look (m_nAppLook) in CEpistemotron ctor and header default.
- Replace several UINT/UINT usages with UINT_PTR/INT_PTR and add static_cast<int> where needed to avoid size/warning issues (EpistemotronView, OutputWnd, Science/Universe).
- Fix TRACE format specifier for timer id.
- Adjust MainFrm toolbar/menu image loading and use int constants for toolbar id ranges.
- Expand and reorganize Epistemotron.rc with full menus, toolbars, strings, icons, bitmaps and include manifest reference; deprecate SimConfigDlg.rc (moved into Resource.rc) and add a placeholder file for reference.
- Add res/app.manifest (visual styles / common controls and compatibility entries).
- Update project file (Epistemotron.vcxproj) to include SimConfigDlg sources and the resource compile entry.
- Add SimConfigDlg header changes (OnOK declared) and stop exposing ConfigureSimulation/OnSimulationTimer in the view header.
- Resource.h and Resource.rc small adjustments (string id, includes) and avoid redefining MFC window IDs.

Overall: this changeset modernizes resources, hardens integer/ptr usage for cross-platform/64-bit builds, and registers the new dialog/manifest in the project.

+++ Science/Universe.h
@@ -29,7 +29,7 @@ public:
 	Universe* GenerateSimulationStep(int p_iStepSize);
 	Mass& GetAt(int p_iPos) { return m_arrMasses.GetAt(p_iPos); }
 	const Mass& GetAt(int p_iPos) const { return m_arrMasses.GetAt(p_iPos); }
-	int GetMassCount() const { return m_arrMasses.GetSize(); }
+	int GetMassCount() const { return static_cast<int>(m_arrMasses.GetSize()); }
 	void ExportPPM(int p_iWidth = 800, int p_iHeight = 600);
 
 	// Optimized symplectic Euler integration for better energy conservation
